+++ docs/README.md
@@ -1,4 +1,4 @@
-# Monique User Guide
+# monique.monique User Guide
 
 // Update the title above to match the actual product name
