+++ src/main/java/monique/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package monique.command;
+
+import monique.exception.DeleteException;
+import monique.storage.Storage;
+import monique.task.Task;
+import monique.tasklist.TaskList;
+import monique.ui.Ui;
+
 public class DeleteCommand extends Command {
 
     private final int taskNum;
@@ -8,7 +16,7 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DeleteException{
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DeleteException {
         if (this.taskNum > tasks.getNumItems() -1 || this.taskNum < 0){
             throw new DeleteException();
         }
