+++ src/main/java/monique/command/ListCommand.java
@@ -1,6 +1,12 @@
+package monique.command;
+
+import monique.storage.Storage;
+import monique.tasklist.TaskList;
+import monique.ui.Ui;
+
 import java.util.stream.IntStream;
 
-public class ListCommand extends Command{
+public class ListCommand extends Command {
 
     public ListCommand(){
         super();
