+++ src/main/java/monique/command/MarkCommand.java
@@ -1,4 +1,11 @@
-public class MarkCommand extends Command{
+package monique.command;
+
+import monique.exception.MarkException;
+import monique.storage.Storage;
+import monique.tasklist.TaskList;
+import monique.ui.Ui;
+
+public class MarkCommand extends Command {
     private final int taskNum;
 
     public MarkCommand(int taskNum){
@@ -7,7 +14,7 @@ public class MarkCommand extends Command{
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws MarkException{
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws MarkException {
         if (this.taskNum > tasks.getNumItems()-1 || this.taskNum < 0) {
             throw new MarkException();
         }
