+++ src/main/java/monique/command/UnmarkCommand.java
@@ -1,4 +1,11 @@
-public class UnmarkCommand extends Command{
+package monique.command;
+
+import monique.exception.MarkException;
+import monique.storage.Storage;
+import monique.tasklist.TaskList;
+import monique.ui.Ui;
+
+public class UnmarkCommand extends Command {
     private final int taskNum;
 
     public UnmarkCommand(int taskNum){
@@ -7,7 +14,7 @@ public class UnmarkCommand extends Command{
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws MarkException{
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws MarkException {
         if (this.taskNum > tasks.getNumItems()-1 || this.taskNum < 0) {
             throw new MarkException();
         }
