+++ src/main/java/monique/exception/MarkException.java
@@ -1,4 +1,6 @@
-public class MarkException extends MoniqueException{
+package monique.exception;
+
+public class MarkException extends MoniqueException {
     public MarkException(){super("Mark Exception: Item Number is not allowed");}
     @Override
     public void advice(){
