+++ src/main/java/monique/task/Event.java
@@ -1,3 +1,7 @@
+package monique.task;
+
+import monique.exception.IllegalDateFormatException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
